\name{erfe-package}
\alias{erfe-package}
\alias{erfe}
\docType{package}
\title{
\packageTitle{erfe}
}
\description{
\packageDescription{erfe}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{erfe}
\packageIndices{erfe}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{erfe}

Maintainer: \packageMaintainer{erfe}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
